新增游戏实时状态栏。

+++ com/akira/undeadwave/core/shop/item/ConsumableShopItem.java
@@ -23,7 +23,7 @@ public class ConsumableShopItem extends ShopItem {
         return consumableItem.getType().getDisplayName();
     }
 
-    protected double getCost() {
+    protected int getCost() {
         return consumableItem.getType().getCost();
     }
 }
